Factor Ident escaping from python code generation
